+++ src/delimited_index_connection.cc
@@ -117,9 +117,6 @@ delimited_index_connection::delimited_index_connection(
     }
   }
 
-  // Check for windows newlines
-  windows_newlines_ = first_nl > 0 && buf[i][first_nl - 1] == '\r';
-
   std::unique_ptr<RProgress::RProgress> pb = nullptr;
   if (progress_) {
     pb = std::unique_ptr<RProgress::RProgress>(
@@ -251,11 +248,7 @@ delimited_index_connection::delimited_index_connection(
       idx_[0].push_back(file_size);
       ++columns_;
     } else {
-      if (windows_newlines_) {
-        idx_[1].push_back(file_size + 1);
-      } else {
-        idx_[1].push_back(file_size);
-      }
+      idx_[1].push_back(file_size);
     }
   }